requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
arxiv==2.2.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.4
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
ddgs==9.6.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
duckduckgo_search==4.1.1
durationpy==0.10
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.12.0
feedparser==6.0.12
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
fpdf2==2.8.4
frozenlist==1.7.0
fsspec==2025.7.0
geojson==3.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
gpt4all==2.8.2
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
InstructorEmbedding==1.0.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jaraco.context==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-google-genai==2.1.9
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.13
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
motor==3.7.1
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.99.5
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.14.1
pyowm==3.5.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
selenium==4.35.0
sentence-transformers==5.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
stack-data==0.6.3
streamlit==1.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.11
unstructured-client==0.42.2
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.37.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weaviate-client==4.16.10
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.3
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.5
xmltodict==1.0.2
yarl==1.20.1
youtube-search==2.1.2
zipp==3.23.0
zstandard==0.23.0